refactor: 프로필 업데이트 영역 분리
